+++ ESFramework/Core/ESMacros.h
@@ -59,4 +59,12 @@
 #define ESLock(lock)    dispatch_semaphore_wait(lock, DISPATCH_TIME_FOREVER)
 #define ESUnlock(lock)  dispatch_semaphore_signal(lock)
 
+/**
+ * Assert that the current code is running on the main thread.
+ */
+#define ESAssertOnMainThread() NSAssert( \
+    es_dispatch_is_queue(dispatch_get_main_queue()), \
+    @"%s should be called on the main thread.", __PRETTY_FUNCTION__ \
+)
+
 #endif /* ESFramework_ESMacros_h */